# Server/src/documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "handyHub",
    "description": "**CHINGU Geckos Team 08**\n\nEnables working with the handyHub API, giving you the customers, suppliers, jobs and  proposals",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/alcb1310/handyHub/0.1.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "security": [
    {
      "password": [
        "read",
        "write"
      ]
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets a list of the users of the API",
        "description": "Retrieves a list of all of the users whether its a supplier of a customer\n- ***supplier*** This required parameter specifies if we are looking for a suplier(true) or a customer(false)",
        "parameters": [
          {
            "name": "supplier",
            "in": "query",
            "description": "If true it will query for all the suppliers and if false all the customers",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Returns the specified number of results",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Represents the page number of the results",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Creates a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "406": {
            "description": "Trades uuid or cities uuid does not exist"
          },
          "409": {
            "description": "Conflict - User already exists"
          }
        },
        "security": []
      }
    },
    "/users/{uuid}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieves an specific user",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Updates a specific user",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "406": {
            "description": "Trades uuid or cities uuid does not exist"
          },
          "409": {
            "description": "Conflict - User already exists"
          }
        }
      }
    },
    "/trades": {
      "get": {
        "tags": [
          "Trades"
        ],
        "summary": "List all of the available trades",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradesResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Creates a new trade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Trade already exists"
          }
        }
      }
    },
    "/trades/{uuid}": {
      "get": {
        "tags": [
          "Trades"
        ],
        "summary": "Retrieves a specifc trade",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trade not found"
          }
        }
      },
      "put": {
        "tags": [
          "Trades"
        ],
        "summary": "Updates a specific trade",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trade not found"
          },
          "409": {
            "description": "Conflict - Trade already exists"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CitiesResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "TradesCreate": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "TradesResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "is_supplier": {
            "type": "boolean"
          },
          "phone": {
            "pattern": "^\\d{3}-\\d{3}-\\d{4}$",
            "type": "string",
            "format": "phone"
          },
          "trades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TradesResponse"
            }
          },
          "cities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CitiesResponse"
            }
          }
        }
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "is_supplier": {
            "type": "boolean"
          },
          "phone": {
            "pattern": "^\\d{3}-\\d{3}-\\d{4}$",
            "type": "string",
            "format": "phone"
          },
          "trades": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "cities": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "password": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "http://example.com/oauth/token",
            "scopes": {
              "write": "allows modifying resources",
              "read": "allows reading resources"
            }
          }
        }
      }
    }
  }
}